requirements.txt
django==1.8.5
djangorestframework==3.3.1
django-rest-swagger==0.3.4
whenpy
requests
django-jet
